# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Blogware RESTful API",
...
        "version": "1.1.2",
        "contact": {
            "name": "Blogware Support",
            "email": "alanmoehammad@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "x-logo": {
            "url": "https://blogware.site/public/files/pictures/scriptlog-1200x630.jpg"
        }
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Relative to the host serving this specification. The spec itself is served from /api/v1/openapi.json, so this resolves correctly on any install with no URL substitution needed. Human-readable examples elsewhere in this document use http://blogware.site as the example host."
        }
    ],
...
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "List published posts",
                "description": "Retrieves a paginated list of published blog posts.\nThis endpoint is public and does not require authentication.\n",
                "operationId": "listPosts",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortOrder"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "description": "Creates a new blog post.\nRequires authentication with administrator, editor, or author permissions.\n",
                "operationId": "createPost",
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a single post",
                "description": "Retrieves a single blog post by its ID.\nThis endpoint is public and does not require authentication.\n",
                "operationId": "getPost",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PostId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a post",
                "description": "Updates an existing blog post.\nRequires authentication with administrator or editor permissions.\n",
                "operationId": "updatePost",
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PostId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post",
                "description": "Deletes a blog post.\nRequires authentication with administrator permissions.\n",
                "operationId": "deletePost",
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PostId"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deleted successfully (no content)"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Posts"
                ],
                "summary": "Partially update a post",
                "description": "Partially updates an existing blog post. Only send the fields you want to change.\nRequires authentication with administrator or editor permissions.\n",
                "operationId": "patchPost",
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PostId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
        "/categories/{id}/posts": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get posts in a category",
                "description": "Retrieves posts belonging to a specific category.\nThis endpoint is public and does not require authentication.\n",
                "operationId": "getCategoryPosts",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/CategoryId"
                    },
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortOrder"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryPostsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
        "/search/posts": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search posts only",
                "description": "Searches only posts for the given keyword (`type` is forced to\n`posts` by the controller).\nThis endpoint is public and does not require authentication.\nSame honeypot (`403`) and endpoint-local rate limit (`429`) as\n`GET /search` apply.\n",
                "operationId": "searchPosts",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search keyword",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2,
                            "maxLength": 100
                        }
                    },
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "API key authentication. Pass your API key in the X-API-Key header.\n"
            },
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Bearer token authentication. Pass your bearer token in the Authorization header.\n"
            }
        },
        "parameters": {
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Page number for pagination",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1,
                    "example": 1
                }
            },
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Number of items per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "default": 10,
                    "example": 10
                }
            },
            "SortBy": {
                "name": "sort_by",
                "in": "query",
                "description": "Field to sort by",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "ID",
                    "enum": [
                        "ID",
                        "post_date",
                        "post_modified",
                        "post_title",
                        "topic_title",
                        "comment_date"
                    ]
                }
            },
            "SortOrder": {
                "name": "sort_order",
                "in": "query",
                "description": "Sort direction",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "ASC",
                        "DESC"
                    ],
                    "default": "DESC"
                }
            },
            "PostId": {
                "name": "id",
                "in": "path",
                "description": "Post ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            },
            "CategoryId": {
                "name": "id",
                "in": "path",
                "description": "Category ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            },
...
        },
        "schemas": {
...
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object"
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
...
            "PostCreate": {
                "type": "object",
                "required": [
                    "post_title",
                    "post_content"
                ],
                "properties": {
                    "post_title": {
                        "type": "string",
                        "maxLength": 200,
                        "example": "My First Blog Post"
                    },
                    "post_content": {
                        "type": "string",
                        "example": "Full post content here..."
                    },
                    "post_summary": {
                        "type": "string",
                        "maxLength": 320,
                        "nullable": true,
                        "example": "Post summary"
                    },
                    "post_status": {
                        "type": "string",
                        "enum": [
                            "publish",
                            "draft"
                        ],
                        "default": "draft"
                    },
                    "post_visibility": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private",
                            "protected"
                        ],
                        "default": "public"
                    },
                    "post_tags": {
                        "type": "string",
                        "nullable": true,
                        "example": "php, api, rest"
                    },
                    "comment_status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "closed"
                        ],
                        "default": "open"
                    },
                    "topics": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                }
            },
            "PostUpdate": {
                "type": "object",
                "properties": {
                    "post_title": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "post_content": {
                        "type": "string"
                    },
                    "post_summary": {
                        "type": "string",
                        "maxLength": 320,
                        "nullable": true
                    },
                    "post_status": {
                        "type": "string",
                        "enum": [
                            "publish",
                            "draft",
                            "pending",
                            "private"
                        ]
                    },
                    "post_visibility": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private",
                            "protected"
                        ]
                    },
                    "post_tags": {
                        "type": "string",
                        "nullable": true
                    },
                    "comment_status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "closed"
                        ]
                    },
                    "topics": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                }
            },
            "PostListResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        }
                    }
                ]
            },
            "PostDetailResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "data": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Post"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "topics": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/components/schemas/CategorySummary"
                                                }
                                            },
                                            "featured_image": {
                                                "type": "string",
                                                "nullable": true,
                                                "example": "http://blogware.site/public/files/pictures/featured.jpg"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                ]
            },
            "PostResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "data": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                ]
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "Bad Request - Invalid parameters",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized - Authentication required",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden - Insufficient permissions",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Not Found - Resource does not exist",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "UnprocessableEntity": {
                "description": "Unprocessable Entity - Validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "Too Many Requests - Rate limit exceeded",
                "headers": {
                    "Retry-After": {
                        "description": "Seconds to wait before retrying",
                        "schema": {
                            "type": "integer",
                            "example": 60
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
        }
    },
...
}